pbio: don't use uint16_t for `now`
it doesn't actuall affect the firmware size (except for in the motor rate ring buffer - so keeping uint16_t there)

+++ lib/pbio/include/pbdrv/ioport.h
@@ -19,7 +19,7 @@ void _pbdrv_ioport_init(void);
  * Do periodic background tasks. This should be called very 1ms.
  * @param [in] now  the current time in millisecond ticks
  */
-void _pbdrv_ioport_poll(uint16_t now);
+void _pbdrv_ioport_poll(uint32_t now);
 
 /** @endcond */
